Fixed old bugs from http://wiki.hping.org/20 : * bad checksum computation. * wrong packet size computation prevented to send max sized packets. * wrong >= prevented to send MTU sized packets. * don't use random IP IDs for fragments of the same packet but use the same ID instead.

+++ cksum.c
@@ -27,7 +27,7 @@ __u16 cksum(__u16 *buf, int nbytes)
 
 	if (nbytes == 1) {
 		oddbyte = 0;
-		*((__u16 *) &oddbyte) = *(__u16 *) buf;
+		*((__u16 *) &oddbyte) = *(__u8 *) buf;
 		sum += oddbyte;
 	}
